Modeling commit 8

+++ Modeling/Modeling/IDraw.cs
@@ -10,9 +10,9 @@ namespace Modeling
     interface IDraw
     {
         void SystemСoordinate(PictureBox pictureBox1, Point coordinateZero);
-        void DrawLine(Point coordinateZero, double zoom, Point startLine, Point endLine);
-        void DrawArcClockwise(Point coordinateZero, double zoom, double radius, Point startPoint, Point endPoint);
-        void DrawArcCounterclockwise(Point coordinateZero, double zoom, double radius, Point startPoint, Point endPoint);
-        void DrawСontour(Point coordinateZero,double zoom);
+        void DrawLine(Point coordinateZero, float zoom, Point startLine, Point endLine);
+        void DrawArcClockwise(Point coordinateZero, float zoom, float radius, Point startPoint, Point endPoint);
+        void DrawArcCounterclockwise(Point coordinateZero, float zoom, float radius, Point startPoint, Point endPoint);
+        void DrawСontour(Point coordinateZero,float zoom);
     }
 }
